feat: support remote sse server
