+++ package.json
@@ -7,7 +7,9 @@
   "scripts": {
     "build": "tsc",
     "dev": "tsx src/index.ts",
+    "dev:remote": "tsx src/remote.ts",
     "start": "node dist/index.js",
+    "start:remote": "node dist/remote.js",
     "test": "npx tsx scripts/test-queries.ts",
     "test:jest": "jest",
     "test:watch": "jest --watch",
